second commit

+++ MovieTicketBooking/src/main/java/com/capg/mtb/repository/IMovieRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.capg.mtb.model.Movie;
 
-public interface IMovieRepository extends JpaRepository<Movie,Integer>{
+public interface IMovieRepository extends JpaRepository<Movie,Long>{
 
 }
